requirements.txt
Flask
groq
pytesseract
Pillow
PyPDF2
python-dotenv
python-docx